+++ base/models/tutor.py
@@ -116,3 +116,8 @@ def search(**criterias):
         for name in full_name.split():
             queryset = queryset.filter(Q(person__first_name__icontains=name) | Q(person__last_name__icontains=name))
     return queryset.distinct().select_related("person")
+
+
+def find_all_summary_responsibles_by_learning_unit_year(a_learning_unit_year):
+    return list(Tutor.objects.filter(attribution__learning_unit_year=a_learning_unit_year,
+                                     attribution__summary_responsible=True))
