+++ base/tests/models/test_tutor.py
@@ -27,13 +27,14 @@ from django.contrib.admin.sites import AdminSite
 from django.contrib.auth.models import Group
 from django.contrib.messages.api import get_messages
 from django.contrib.messages.storage.fallback import FallbackStorage
-
-from base.models import tutor
 from django.test import TestCase
 
+from base.models import tutor
 from base.tests.factories.tutor import TutorFactory
 from base.tests.factories.person import PersonFactory
 from base.tests.factories.user import UserFactory
+from attribution.tests.models import test_attribution
+from base.tests.factories.learning_unit_year import LearningUnitYearFactory
 
 
 class TestTutor(TestCase):
@@ -43,6 +44,11 @@ class TestTutor(TestCase):
         self.tutor = TutorFactory(person=self.person)
         TutorFactory()  # Create fake Tutor
         TutorFactory()  # Create fake Tutor
+        self.learning_unit_year = LearningUnitYearFactory()
+        self.attribution = test_attribution.create_attribution(tutor=self.tutor,
+                                                               learning_unit_year=self.learning_unit_year,
+                                                               score_responsible=False,
+                                                               summary_responsible=True)
 
     def test_find_by_person(self):
         self.assertEqual(self.tutor, tutor.find_by_person(self.person))
@@ -73,9 +79,13 @@ class TestTutor(TestCase):
     def test_find_by_id_wrong_id(self):
         self.assertIsNone(tutor.find_by_id(-1))
 
+    def test_find_all_summary_responsibles_by_learning_unit_year(self):
+        responsibles = tutor.find_all_summary_responsibles_by_learning_unit_year(self.learning_unit_year)
+        self.assertCountEqual(responsibles, [self.tutor])
+
 
 class MockRequest:
-    COOKIES={}
+    COOKIES = {}
 
 
 class MockSuperUser:
